Move preDeployValidation steps in app to function spec
