+++ packages/app/src/cli/models/app/app.test.ts
@@ -7,24 +7,17 @@ import {
   getUIExtensionRendererVersion,
   isCurrentAppSchema,
   isLegacyAppSchema,
-  validateExtensionsHandlesInCollection,
-  validateFunctionExtensionsWithUiHandle,
 } from './app.js'
 import {
   DEFAULT_CONFIG,
   testApp,
   testUIExtension,
-  testFunctionExtension,
   testWebhookExtensions,
-  testEditorExtensionCollection,
   testAppAccessConfigExtension,
   testAppHomeConfigExtension,
   testAppProxyConfigExtension,
 } from './app.test-data.js'
-import {ExtensionInstance} from '../extensions/extension-instance.js'
-import {FunctionConfigType} from '../extensions/specifications/function.js'
 import {WebhooksConfig} from '../extensions/specifications/types/app_config_webhook.js'
-import {EditorExtensionCollectionType} from '../extensions/specifications/editor_extension_collection.js'
 import {ApplicationURLs} from '../../services/dev/urls.js'
 import {describe, expect, test} from 'vitest'
 import {inTemporaryDirectory, mkdir, writeFile} from '@shopify/cli-kit/node/fs'
@@ -229,163 +222,6 @@ describe('getAppScopesArray', () => {
   test('returns the access_scopes.scopes key when schema is current', () => {
     const config = {...DEFAULT_CONFIG, access_scopes: {scopes: 'read_themes, read_order ,write_products'}}
     expect(getAppScopesArray(config)).toEqual(['read_themes', 'read_order', 'write_products'])
-  })
-})
-
-describe('validateFunctionExtensionsWithUiHandle', () => {
-  const generateFunctionConfig = ({type, uiHandle}: {type?: string; uiHandle?: string}): FunctionConfigType => ({
-    description: 'description',
-    build: {
-      command: 'echo "hello world"',
-      wasm_opt: true,
-    },
-    api_version: '2022-07',
-    configuration_ui: true,
-    metafields: [],
-    name: 'test function extension',
-    type: type || 'product_discounts',
-    ui: {
-      handle: uiHandle || 'test-ui-handle',
-    },
-  })
-
-  describe('returns errors when app configuration is invalid', () => {
-    test("when a function's ui handle does not match any local ui extension", async () => {
-      // Given
-      const validFunctionWithUiExtension = await testFunctionExtension({
-        config: generateFunctionConfig({uiHandle: 'test-ui-extension'}),
-      })
-      const allExtensions: ExtensionInstance[] = [validFunctionWithUiExtension]
-      const app = await testApp({
-        allExtensions,
-      })
-
-      // When
-      const expectedErrors = [
-        "[test function extension] - Local app must contain a ui_extension with handle 'test-ui-extension'",
-      ]
-      const result = validateFunctionExtensionsWithUiHandle([validFunctionWithUiExtension], app.allExtensions)
-
-      // Then
-      expect(result).toStrictEqual(expectedErrors)
-    })
-
-    test('returns error when a functions matching extension, and not of type ui extension', async () => {
-      // Given
-      const functionWithUiHandle = await testFunctionExtension({
-        config: {
-          ...generateFunctionConfig({type: 'product_discounts', uiHandle: 'product_discounts-test'}),
-          handle: 'product_discounts',
-        },
-      })
-      const functionWithMatchingHandle = await testFunctionExtension({
-        config: {
-          ...generateFunctionConfig({type: 'product_discounts'}),
-          handle: 'product-discounts-test',
-        },
-      })
-
-      const allExtensions: ExtensionInstance[] = [functionWithUiHandle, functionWithMatchingHandle]
-      const app = await testApp({
-        allExtensions,
-      })
-
-      // When
-      const expectedErrors = [
-        "[test function extension] - Local app must contain a ui_extension with handle 'product_discounts-test'",
-      ]
-      const result = validateFunctionExtensionsWithUiHandle([functionWithUiHandle], app.allExtensions)
-
-      // Then
-      expect(result).toStrictEqual(expectedErrors)
-    })
-
-    test('returns errors for editor extension collection', async () => {
-      // Given
-      const configuration = {
-        name: 'Order summary',
-        handle: 'order-summary-collection',
-        includes: ['handle1', 'product-discounts-test', 'admin-extension', 'customer-account-extension'],
-      }
-      const editorExtensionCollection = (await testEditorExtensionCollection({
-        configuration,
-      })) as ExtensionInstance<EditorExtensionCollectionType>
-
-      const orderDiscountFunction = await testFunctionExtension({
-        config: {
-          ...generateFunctionConfig({type: 'product_discounts'}),
-          handle: 'product-discounts-test',
-        },
-      })
-
-      const adminUiExtension = await testUIExtension({
-        type: 'ui_extension',
-        configuration: {
-          type: 'ui_extension',
-          handle: 'admin-extension',
-          extension_points: [
-            {
-              target: 'admin.customers.segmentation-templates.render',
-              module: './src/ExtensionPointA.js',
-            },
-          ],
-        },
-      })
-
-      const customerAccountUiExtension = await testUIExtension({
-        type: 'ui_extension',
-        configuration: {
-          type: 'ui_extension',
-          handle: 'customer-account-extension',
-          extension_points: [
-            {
-              target: 'customer-account.order-index.block.render',
-              module: './src/ExtensionPointB.js',
-            },
-          ],
-        },
-      })
-
-      const allExtensions: ExtensionInstance[] = [
-        editorExtensionCollection,
-        orderDiscountFunction,
-        adminUiExtension,
-        customerAccountUiExtension,
-      ]
-
-      const app = await testApp({
-        allExtensions,
-      })
-
-      // When
-      const expectedErrors = [
-        "[order-summary-collection] editor extension collection: Add extension with handle 'handle1' to local app. Local app must include extension with handle 'handle1'.",
-        "[order-summary-collection] editor extension collection: Remove extension of type 'function' from this collection. This extension type is not supported in collections.",
-        "[order-summary-collection] editor extension collection: Remove extension 'admin-extension' with target 'admin.customers.segmentation-templates.render' from this collection. This extension target is not supported in collections.",
-      ]
-      const result = validateExtensionsHandlesInCollection([editorExtensionCollection], app.allExtensions)
-
-      // Then
-      expect(result).toStrictEqual(expectedErrors)
-    })
-
-    test('returns undefined when validation passes', async () => {
-      // Given
-      const validUiExtension = await testUIExtension({type: 'ui_extension'})
-      const validFunctionWithUiExtension = await testFunctionExtension({
-        config: generateFunctionConfig({uiHandle: 'test-ui-extension'}),
-      })
-      const allExtensions: ExtensionInstance[] = [validUiExtension, validFunctionWithUiExtension]
-      const app = await testApp({
-        allExtensions,
-      })
-
-      // When
-      const result = validateFunctionExtensionsWithUiHandle([validFunctionWithUiExtension], app.allExtensions)
-
-      // Then
-      expect(result).toBeUndefined()
-    })
   })
 })
